+++ examples/physics_rapier_settings.rs
@@ -64,8 +64,8 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             c.insert((
                 TilemapAnchor::Center,
                 TiledPhysicsSettings::<TiledPhysicsRapierBackend> {
-                    objects_layer_filter: TiledName::None,
-                    tiles_objects_filter: TiledName::Names(vec![String::from("collision")]),
+                    objects_layer_filter: TiledFilter::None,
+                    tiles_objects_filter: TiledFilter::Names(vec![String::from("collision")]),
                     ..default()
                 },
             ));
@@ -79,8 +79,8 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
             c.insert((
                 TilemapAnchor::Center,
                 TiledPhysicsSettings::<TiledPhysicsRapierBackend> {
-                    objects_layer_filter: TiledName::All,
-                    tiles_objects_filter: TiledName::None,
+                    objects_layer_filter: TiledFilter::All,
+                    tiles_objects_filter: TiledFilter::None,
                     ..default()
                 },
             ));